requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
psycopg2-binary==2.9.3
redis==4.3.4
python-binance==1.0.16
prometheus-client==0.14.1
pytest==7.1.2
pytest-asyncio==0.18.3
pytest-mock==3.10.0
pytest-cov==4.0.0
python-dotenv==0.19.2
numpy==1.21.0
aiohttp==3.8.1
asyncio==3.4.3
psutil==5.9.5